makefile: rename $CPP to $CXX, to avoid clashing with C preprocessor
CPP is traditionally used to provide the location of the C preprocessor in
the environment; using it to designate the C++ compiler breaks environments
which assume CPP points to the preprocesor (e.g., this makefile wouldn't
compile withc conda-build, that sets CPP).

Renaming CPP to CXX (the same convention used in find_orb makefiles).